FEATURE: adds second search bar to Force
